Change from using double to BigDecimal for the price.

+++ ApplicationServerLiberty/src/main/java/Application/Entities/Stock.java
@@ -1,17 +1,18 @@
 package Application.Entities;
 
 import jakarta.enterprise.context.RequestScoped;
+import java.math.BigDecimal;
 
 @RequestScoped
 public class Stock {
     private String symbol;
     private String company;
-    private double price;
+    private BigDecimal price;
     private String country;
 
     public Stock() {}
 
-    public Stock(String symbol, String company, double price, String country) {
+    public Stock(String symbol, String company, BigDecimal price, String country) {
         this.symbol = symbol;
         this.company = company;
         this.price = price;
@@ -22,7 +23,7 @@ public class Stock {
 
     public String getCompany() {return company;}
 
-    public double getPrice() { return price; }
+    public BigDecimal getPrice() { return price; }
 
     public String getCountry() { return country; }
 }
